frontend ui installed

+++ backend/controllers/product.controller.js
@@ -70,6 +70,9 @@ export const deleteProduct = async (req, res) => {
   const { id } = req.params; // Extract the ID from the URL parameter
   // console.log("id:", id);
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ success: false, message: "Invalid Product Id" });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Product deleted" });
